Using the compiler that comes with Arduino 1.0.5, it complains that the token in the async struct cannot have a size of zero.
I fixed it, since it's the end of the struct, this is okay.

Patch 2: Updated with respect to Sachin's comment review.

Patch 3: When building in debug mode, the compiler that
comes with Arduino 1.0.5 complains about the tokens defined
in pdu.h as well. In this patch as well, also removed errors
found when building in debug. Rebased on Sachins changeset.

Change-Id: I568481562af2a614624608851def5bf5aa68287a