Patch 1: Arduino pdu malloc fix.
Patch 2: Rebase

Patch 3: Changed implementation of fix to simply change size of
COAP_MAX_PDU_SIZE to 320 for Arduino only. This will stay the same
going forth, but there is an uncaught case that can only be fixed
in the next sprint(as there is not enough time)-- We must implement
blockwise transfer so that a PDU larger than 320 is properly
handled (i.e. broken up into chunks of 320 or smaller). A Rally task
has already been created for Sprint 8 for this.

Patch 4: Meant to use "else" instead of "elif".

Change-Id: I4df20d28c821f6fc2c98a6bf3672331264978be5