Fixing a few warnings that we see coming out of the libcoap bits.h file.
Essentially, the result of the bit-shift on a uint8_t is always upconverted to a signed int despite it being impossible for it to be anything greater than 5 bits long (8 bits, shifted 3 more).  We cast to unsigned size_t to prevent the warning and to prevent unsigned vs signed comparison.

Change-Id: I47d05611e75e55240b55b800c9370dbc019aedd2