Makefile changes req'd for a global build script. Also, global build script implementation "buildScript.mk." Updated base README to reflect the options and requirements for running this build script.
Patch 2: Edited makefiles so that developers CAN continue to use the build
processes they're used to. Also added 'all_dev' target that will build as
much of everything the build server can (including OCICUC) to prevet
regression in the build process as much as possible with a full-blown
build script.

Patch 3: Had to edit oc_logger makefile to support the new directory output
structure from the edits of this changeset. Also took care of issues
that Erich pointed out.

Patch 4: Added notes to README to state what is NOT being built by the
current iteration of buildScript.mk. Also edited lines of README that Doug
had mentioned needed updated in Patch 3.

Patch 5: Edited mistype in README and edited csdk/makefile as suggested by
Sachin.

Note: Arduino Due is not a part of the build script process.  This changeset
needs to be verified on the build server and put in place as the sole build
process. Once this is done, Support for building the Arduino Due will go in
place. Also, Ethernet Vs. Wifi build support will go in place at that time.
Right now everything is Ethernet.

Change-Id: Ie9938d9f3069aff1858141565dc66b4e2ece9e98