Added logic to notify all observers when a resource has been deleted by OCDeleteResource API or when OCStop() is called. This will effectively tell all observers that "this" resource is no longer available. This information can be used by a client to infer that the server is going down.
If any more changes are req'q to encompass the new "Client able to delete
a resource" user story, a new changeset will be generated.

To come: A changeset containing a set (server and client) of examples
using this functionality.

Patch 2: Thuyen rebased on top of mine and for some reason needed to change
the author of my changeset, but doesn't appear to have made any changes to
my changeset other than rebasing it onto master.

Patch 3: Rebased on to newest master and made changes req' by Sachin and
Sudarshan in Patch 1 comments.

Change-Id: I343a5589c2988c45da95ba6d15685c21fa4704b6