Move base makefiles from source directories to build directory.
This commit was originally authored by Ashok Channa, and modified with
subsequent commits by Bill Dieter to meet Samsung's request to move
makefiles out of their locations in the source directories into the
build directories to match the way they build for Tizen.

The changes have all been squashed together to avoid a commit that does
not compile.

In the process of changing directories, the top level "clean" target is
updated to loop through directories and only clean those that exist, to
prevent failures due to trying to clean non-existent build directories.

Change-Id: Iff59b2f322ee75f517ff870557b17931d91f309f
Signed-off-by: William R. Dieter <william.r.dieter@intel.com>