Updated libcoap to send messages using secure port
Whenever a REST request is received by OIC Stack on secure
port, it needs to keep track of this fact for two reasons:
1) It allows upper layer to grant/deny access to secure resource
if the request was unauthenticated.
2) It also allows OC Stack to encrypt the response for a secure
request.
Since existing code was using a mechanism of send_flag to keep
track of various send mechanisms, it was modified to a bit mask
flag.

Change-Id: I293d4378a1e90234daf2553b6c6defdb6a3abac5
Signed-off-by: Sachin Agrawal <sachin.agrawal@intel.com>