Added support for multicast presence
Added OCMulticastNode data structure to store the nonce information of the server send
presence notifications. When a new notification comes to the client, client checks if there
is a multicast node, and then checks to see if the nonce is same or not. If same nonce, no call
back to client application is made. If nonce is different a call to client application is made.

Signed-off-by: Sashi Penta <sashi.kumar.penta@intel.com>
Change-Id: Ia3fe163e7a304f499272f9ffa8a22a359f123d98