Implement sending host port number for secure resources
Secure resources are hosted on a separate IP port. All the
discovery responses are send from unicast unsecure port from
resource server. Currently, when a client receives discovery response
it 'assumes' that the sending port is the port hosting the resource
too. This 'assumption' is NOT valid for secure resources. Therefore,
there is a need to send 'secure port info' for secure resources in
JSON discovery response.

Change-Id: I54fd828a15913ec8be990d040e1190a88be43353
Signed-off-by: Sachin Agrawal <sachin.agrawal@intel.com>