Fix the infinite PRESENCE_STOPPED message
When a client to subscirbes to a server, and when the server process is
exited, client is getting was getting infinte PRESENCE_STOPPED messages.
This commit fixes this issue. Introduced a new stack result
type : PRESENCE_TIMEOUT to indicate timeout of the server.
PRESENCE_STOPPED is used for graceful stopping of the server.

Change-Id: I4e122176f8f7ea8e04d98dea02d22432b4c4555a
Signed-off-by: Sashi Penta <sashi.kumar.penta@intel.com>