Implement slow response feature.
This feature allows a resource entity handler to provide a "slow response"
to a client request.  That is, if a request is received but the server cannot
respond immediately (e.g. response data is not available), the server can
instead respond at a later time.  The same API is used for both immediate
and slow responses.

Updated collection's batch interface to aggregate responses of local resources
contained in the collection.  Note that the aggregate response only applies
to local resources, not remote resources.

This feature introduces a new C API, OCDoResponse and a new C++ API,
OCPlatform::sendResponse.  The C++ modifications also include a change to
the the signature of the applications entity handler.

Change-Id: Ic33ab0c7ff52481bed86e9a77319baa3813dd6cd
Signed-off-by: Doug Hudson <douglas.hudson@intel.com>
Signed-off-by: Yamin Al-Mousa <yamin.s.al-mousa@intel.com>
Signed-off-by: Sashi Penta <sashi.kumar.penta@intel.com>
Signed-off-by: Sachin Agrawal <sachin.agrawal@intel.com>
Signed-off-by: Sudarshan Prasad <sudarshan.prasad@intel.com>
Signed-off-by: Erich Keane <erich.keane@intel.com>