Fixed usages of memset(..0..) to use Calloc/init'er
Our code has a number of cases where we declare a struct then
immediately memset to all 0.  This uses the initializers which
are less prone to failure and are faster due to compiler optimizations.

Additionally, we have a number of cases where we OCMalloc a
struct, then immediately zero it.  This fix uses the new OCCalloc
in order to save unnecessary memset and reduce potential for accidental
error.

In a few of the cases, older compilers didn't handle the {0} initializer
correctly and gave errors (Arduino Mega).  Because of that, a few structs
were reordered to compile for that platform.

Change-Id: I56f29eac68240abea2fb8a69d888a09d61d4488a
Signed-off-by: Erich Keane <erich.keane@intel.com>