Moved all presence functionality into the stack layer from coap layer.
This task was initiated so that observe parameters can be polled from
whatever new(er) transport layer we decide to use in the future and then
specific presence parameters can be determined once we have those observe
parameters. This effectively causes all presence related work to be done
in the stack layer, while observe related parameters are transferred OTA
and presence parameters are simply extrapolated after the fact.

This changeset correlates to Rally User Story: US2195 "As an Iotivity
client and server developer, all presence related functionality in the
OCCoap layer is moved to the stack layer."

Change-Id: Ibb784946fe233e17834136ed96ebed3436372742
Signed-off-by: Joseph Morrow <joseph.l.morrow@intel.com>