Make all multicast coap messages sent as Low_QoS/NON.
When a client in C++ is configured to use HighQoS,
then discovery request goes as CON to a multicast address.
However, responses/ACKs are coming from a unicast address.
This makes it impossible for clients to associate requests
with responses and stop retrying CON pdus.Now, CoAP layer is
making sure that all packets sent to multicast address are NON.

Change-Id: I41f351c5c9ce75fc4846d970514733bef5cdce61
Signed-off-by: Yamin Al-Mousa <yamin.s.al-mousa@intel.com>
Signed-off-by: Sudarshan Prasad <sudarshan.prasad@intel.com>