Download gtest-1.7.0 rather than depending on it being installed
Making the user download and install gtest-1.7.0 is inconvenient for the
user.  Also, we do not have direct access to the build machine to update
how and where gtest is stored.  Downloading to the extlibs directory in
project source tree removes the need for developers to download and build
gtest and removes the need to configure the build machine.

Change-Id: I3b163c09ecdd9f12e8004a69dfa8dbcaf4ae5615
Signed-off-by: William Dieter <william.r.dieter@intel.com>