Fixes for scons builds
1) Remove oic-utilities as a requirement

2) Fail the build if one step fails

Setting this option in the bash script will cause the
script to terminate if any of the commands returns a
with non-zero value.  This is necessary for the build
server to understand when a build is failed.  Without this
flag, the script will exit with the status of the last command,
which in this script is "echo" so it will be 0, regardless if
a scons build step failed.

Ideally we would try each step and then report the failures
at the end, set -e will do the job until that can be developed.

3) Conditionaly enable android build

4) Correct a build issue in occoap.c with typecasting

Change-Id: I2711db6ff0814cff64b170cba20455e8ef52f905
Signed-off-by: David Warburton <david.warburton@intel.com>