Fail the build if one step fails
Setting this option in the bash script will cause the
script to terminate if any of the commands returns a
with non-zero value.  This is necessary for the build
server to understand when a build is failed.  Without this
flag, the script will exit with the status of the last command,
which in this script is "echo" so it will be 0, regardless if
a scons build step failed.

Ideally we would try each step and then report the failures
at the end, set -e will do the job until that can be developed.

Change-Id: Ib6692d434a2e811f61d8dd1bff52766b7511abd7
Signed-off-by: David Warburton <david.warburton@intel.com>