OICSensorBoard: App for Intel Edison Platform.
(Please refer to https://oic-review.01.org/gerrit/#/c/668/8/
for prior review status) 

Build for OICSensorBoard is NOT included in the
standard Iotivity build.

It requires the Yocto cross-toolchain for
a chosen target, and would therefore be
built separately.

It demonstrates Iotivity Server capabilities through
the integration of an add-on breadboard that hosts
temperature, ambient light and LED sensors.
In addition, there is a command-line Iotivity Client
application to test server functionality from a
separate machine. 

The included scons build script expects to build
the client for Linux (tested on Ubuntu 12.04)
and the server to be built via the Yocto
cross-toolchain for Edison.

Brief instructions for constructing this
toolchain are provided in the included README.

Change-Id: I26d12c2334d84a6e0245e0f3689e09a51152fea5
Signed-off-by: Kishen Maloor <kishen.maloor@intel.com>
(cherry picked from commit 7cd6e2214c0eb9a70ac8cb94b38f9dc682a6691e)
Reviewed-on: https://gerrit.iotivity.org/gerrit/182
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sudarshan Prasad <sudarshan.prasad@intel.com>