Android SDK - Initial commit and build system updates
This commit add makes the auto_build process more complete by
adding the following:

1) Auto build now always performs the dependency check
2) Auto build now always attempts to build the Android stack artifacts.
3) Transitioned the Android build process to the v4.9 tool chain
4) Android build now depends on the internal boost dependencies
5) Add an Android compatibility layer for the code that is using
   the C++ 11 basic string functions with is not supported in the
   Android toolchain yet.
6) Added static libraries to many of the shared libraries for the
   Android SDK

NOTE:
a) This commit conflicts with the work that Caiwen was making in the
   C++11 basic_string function compatibility. I will close with him.
b) Expat is needed by the protocol plug in manager for Android.
   This still needs to be closed by release. The PPM build for
   Android under SCONS is currently disabled.

This is the Android SDK client API.
Validation has not been completed. However, this is all new
functionality and does not have any negative impact on any
existing features. Until validation is completed this
functionality should be considered experimental.

This code based currently uses gradle to build the SDK.
It does require that the Android libraries for IoTivity
be built previously using the appropriate SCONS scripts.

This additional patch attempts to transition the boost
build to using git instead of a download in hopes that
the build machine will be happier about it.

Final patch makes Android build optional in auto_build

Another patch to add support for 64-bit hosts

Another patch set to reduce parallel builds to 4

Another patch to add licensing header to two files that we missing
in gerrit review #72

Change-Id: I9695ebf3ea271ec28b69a4fcec509a6c2b7c46a4
Signed-off-by: Patrick Lankswert <patrick.lankswert@intel.com>