Implement [] and iteration for OCRepresentation
Multiple requests have come in to be able to access the contents
of OCRepresentation in an easier to use manner.  This fix allows
OCRepresentation to be indexed using the [] operator and also
be iterated in an STL compliant fashion.

Additionally, this adds more than 40 tests for this functionality


Change-Id: Ibfc4bc61493324e118c4a6b67efc1d66e8b66b04
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/116
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>
(cherry picked from commit 529bb98b884ca4db27b7659f966075a8680fa3c3)
Reviewed-on: https://gerrit.iotivity.org/gerrit/325
Tested-by: Sashi Penta <sashi.kumar.penta@intel.com>