Fixed invalid memory read, removed dead code
Debug.c's coap_print_addr had its functionality completely removed
in the past.  However, this resulted in returning SOMETHING (not zero!)
despite not actually correctly writing to the addr parameter.

Since this parameter is uninitialized when sent to this function,
and the return value is used to determine whether it correctly added
info to the addr, the debug print following this (net.c:972) attempted
to read the uninitialized data, potentially causing memory corruption.

This fix removes the 'dead' code, then returns a zero (failure, nothing
written).

Change-Id: If4f2c131d9b2f585299363f21c3686c25de29104
Signed-off-by: Erich Keane <erich.keane@intel.com>