Fixing repeated EH calls when client retransmits requests
(Bug IOT-239) When a client makes a request to a slow resource
and it does not receive an ACK for that request, it retransmits
the requests which resulted in calling the server's entity handler
several times and multilpe slow responses. When a server
detects a repeated request it should not process it. Also, when
the client receives a slow response it should stop retransmitting
requests.

Change-Id: I101ad3de9b8e0e4990fb096d4a8fe5b9dcebf4b0
Signed-off-by: Yamin Al-Mousa <yamin.s.al-mousa@intel.com>