Fixed InProcServerWrapper to be thread safe
A few of the variables/collections in the server wrapper were not
properly protected against multi-threading.  This fix adds a mutex
to the class to ensure that all accessed items are protected.
Additionally, they are moved out of the root namespace to prevent
polution.

Change-Id: I2eedb9bf9b1960ff7a018a2ca1e15183e42bd9cd
Signed-off-by: Erich Keane <erich.keane@intel.com>
(cherry picked from commit a31834d16272a5b8d3382fb2959676804157c732)
Reviewed-on: https://gerrit.iotivity.org/gerrit/214
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sudarshan Prasad <sudarshan.prasad@intel.com>