Fixed crash on rec'd invalid data.
This issue is a fix for IOT 114.

Previously, a bad series of JSON coming in would cause an exception
to be thrown. this would result in the callback/server crashing, which
is not appropriate behavior. This fix catches and logs any JSON
errors keeping the callback/server from crashing.

NOTE: This changeset is originally from 01.org(changeset 695,
author: Erich Keane).

Change-Id: I1e7d042f243bc1b86771778c308f1187bcda875f
Signed-Off-by:Sudarshan Prasad<sudarshan.prasad@intel.com>
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/103
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>