Improve usability of ActionSet for Group Action Feature.
ActionSet structure is not show directly to developer.
Because there is the definition of ActionSet in GroupManager.h.
So we separate the definition of ActionSet from Groupmanager to ActionSet.h
And it is exposed to developer by one type of header file.

Change-Id: I2ccb93cf70ec42b122aaead4943aed56df161641
Signed-off-by: HyunJun Kim <hyunjun2.kim@samsung.com>
(cherry picked from commit 9520eb3c228275be55a3ee0401a8c590571c3168)
Reviewed-on: https://gerrit.iotivity.org/gerrit/223
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sudarshan Prasad <sudarshan.prasad@intel.com>