Fix issue where logger files aren't built for Arduino on CA Branch.
The logger file has to be built with a C++ compiler so that we can
use Arduino's Serial functions as they are prepared in C++ world. With
a ".c" extension, the CA layer makefiles enforce it to be build with
a C compiler. We are forcing only the logger.c file to built for c++.


Change-Id: If2c17d15ebdb3b9942cb6c8391b4c2f7865b1fcd
Signed-off-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/187
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>
Reviewed-by: Sudarshan Prasad <sudarshan.prasad@intel.com>