Fix errors on Arduino caused by proper initialization.
Initializing structs with "{0}" causes errors on Arduino where the
struct's first member is a non-primitive or something that can not be
properly set to "0" without a cast. I re-arranged the structs such that
the first member of the structs are of a primitive type or are typedef'd
as a primitive type.

Change-Id: Ib7ecced34ea85b1a8c9106c34e1d27890a4cf039
Signed-off-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/304
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>