Clean up TODOs in OCUtilities
The Escape String mechanism is no longer used anywhere, so it is
dead code.  The QueryParam parsing was only used in 1 place and the
initial non 'temp patch' function was all dead code.  This fix removes
both sets of dead code and cleans up the temp-patch code to be a little
more efficient.

Change-Id: I0777f5d5795362195e35bdffc82a6eed55850a91
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/320
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>