Repaired a number of issues discovered while doing a code-review
I was poking through a general code review/testing out a SCA tool
and came up with/discovered a handful of potential bug-filled code.

In a few locations we were referring to a variable BEFORE testing for
validity which has some pretty nasty concequences.  I replaced usages
of strtok with strtok_r in a few places I came across it.  The IP
parsing code was a bit convoluted and unnecessary, so I made it more
clear.

Change-Id: If1157ca33d2fe176a4beac9b3e4b8a8090357751
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/346
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Doug Hudson <douglas.hudson@intel.com>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>