Retrieve build flags from pkg-config only once.
The pkg-config command used to retrieve the build flags for several
libraries in Linux, Tizen and Android builds was added to the C and
C++ command line flags as a shell "backtick" expression, e.g.:

  gcc ... `pkg-config glib-2.0 --cflags` ...

That caused pkg-config to be run for each and every source file
compilation command executed by SCons.  Doing so is an expensive
operation since it causes an additional process to be spawned to read
the same flags for each source file being compiled.

Just have SCons read them and merge them to the build environment
once!

Change-Id: I3120aef7929464c033c8dabf6ec0ca88e5b67f89
Signed-off-by: Ossama Othman <ossama.othman@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/367
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>