Changed OCDOResource's handle param to take NULL
Most calls to OCDoResource were ignoring the handle parameter's
out value, so this call formalizes that a NULL is an acceptable value
for this parameter.  The callers that were using this (as highlighted
by Static Code Analaysis) without using the handle parameter were all
switched to use this new NULL functionality.

Change-Id: I9fc047c080499d1283ffc2cba32addd0f5689c77
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/400
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>