Removed unneeded references to oic-utilities
The oic-utilities repo was used as a temporary solution for holding third
party libraries.  The build system now has a better way to handle them.
This commit removes obsolete instructions to copy the oic-utilities repo
from the scons build instructions.

The resource/csdk/README.txt file still has a reference to oic-utilties,
as do some of the old release notes files.  The csdk README can probably
be deleted when makefiles are eliminated.

Change-Id: Ic5824264835359a59cfd79443ce4e773205eeb07
Signed-off-by: William Dieter <william.r.dieter@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/420
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Caiwen Zhang <caiwen.zhang@intel.com>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>