Fixed InitializeException issues
InitializeException had a number of minor issues that are repaired in
this fix.  Note that the interface of InitializeException changes a bit,
however it seems that no one is actually catching InitializeException
anywhere, so they wouldn't be able to use the old version.

This fix switches InitializeException to inherit from OCException,
permitting us to catch it properly, rather than being a different exception
type.  This results in the reason, code, and 'what' methods working properly
in both cases, rather than differing slightly by behavior.

Change-Id: If41377af1b1c4ef7297ae9644ef2e7e7e4f8b603
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/483
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sakthivel Samidurai <sakthivel.samidurai@intel.com>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>