Removed Dead FormOCResponse Code
The function FormOCResponse was previously only used in one place,
and had no side effects other than allocating memory and filling an
object with a bunch of different values.

The calling site (OCProcessPresence) only refers to the OCResponse
object in this single call, then leaks it immediately after.

This changeset removes the function, the variable, and the call to
this function.

Change-Id: Ic40440fce04343149528fb4d33f42f73da908171
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/494
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>
Reviewed-by: Sudarshan Prasad <sudarshan.prasad@intel.com>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>