Fixed Arduino config setup when 'ALL' is selected as a transport.
Arduino scons config was only structured to handle one transport as a time,
and would fail to configure needed lib imports when 'ALL' was the selected
transport.

This change removes the choice of 'ALL' when Arduino is the selected OS. It
also will log an error and stop the build when more than a single transport
is selected for Arduino builds. This is mainly until resources permit
testing of the code on hardware configurations with multiple transport
shields attached.

Change-Id: I9da97df1db696904581eaccbbbc519b291b0b3ed
Signed-off-by: Jon A. Cruz <jonc@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/507
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>