Removed a single instance of the usage of a "magic number" in the stack.
After walking through the logic in ValidateQuery() its clear that there
was an intention add a second #define, but this must have been over-
looked at that time. There a single #define to avoid usage of a
"magic number" (or a raw value, ie. "2") but it was actually being used
in the wrong spot in the code.

Change-Id: I2ff045005ccacc9822280d93e6b20bfac8df1cbd
Signed-off-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/539
Reviewed-by: Erich Keane <erich.keane@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>