Removed Token ASCII-char limit and length limit
Currently, the stack has a fixed-size for Tokens (CAToken_t)
which is null terminated and limited to the ASCII character
set.  The OIC specification requires that tokens be variable
length up to 8 bytes, not be null terminated, and be the 
full binary set.

This change removes the null termination behavior (bad,
since it is binary data, where the NUL is a valid value),
provides for a variable length token, and removes the previous
limit that held us to the printable ASCII character set.

This issue was captured in IOT-380 in JIRA.

Change-Id: I2159dd3202edeff4025c7e1d0de50360b6521e53
Signed-off-by: jnashok <jn.ashok@samsung.com>
Signed-off-by: Sakthivel Samidurai <sakthivel.samidurai@intel.com>
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/598
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>