Adding base timestamps to linux log output.
This updates logging output on linux to have basic timestamps prepended.

It includes additions of link flag adjustments needed for gcc 4.6.

This currently uses coarse timers to allow for faster response. The
timer type could be switched to normal if testing something more
precise. On some systems tested the resolution was 4ms.

Change-Id: Ief75680bfd17178e4c035ef6a69db00b8c7de73e
Signed-off-by: Jon A. Cruz <jonc@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/622
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>