Add stack logging flag to scons build parameters.
This will allow people to build debug versions of the stack without
actually enabling all the stack debug logs. This is useful (for instance)
when you may want to run unit tests in debug mode (ie. with no optimizat-
ion and full debug symbols), run without stack logging noise, and run
valgrind as the CI (ie. build) system does. The default values of when
logging will be enabled remain the same as before, except the default for
stack logging when building/running the unit tests is disabled. You may
change this by simply adding "LOGGING=1" or "LOGGING=true" to the list
of scons options, however.

Change-Id: I70cedc9143c6015fd75bf4725cf72b6a64b8529e
Signed-off-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/657
Reviewed-by: Erich Keane <erich.keane@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>