Update the use of clock_gettime(2)
clock_gettime(2) is a POSIX feature, but CLOCK_REALTIME_COARSE is a
Linux extension. So instead of using clock_gettime only on Linux, try to
use the POSIX macros instead. This way, we'll use clock_gettime on
systems that support it, falling back to gettimeofday if they don't.

POSIX.1-2008 requires clock_gettime(2) functionality, so up the
_POSIX_C_SOURCE macro requirement to the 2008 version. This will request
the 2008 API if it's available on some systems (QNX is known to not
enable all the API it has available). Older versions of POSIX don't
require the timer support.

Change-Id: I27eaacb532114dd188c4ffff13d3229d285d21cb
Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/677
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Charlie Lenahan <charlie.lenahan@intel.com>
Reviewed-by: Erich Keane <erich.keane@intel.com>