Cleaning previous debug build before running Unit Tests to rebuild debug.
By doing this, it will cause a trigger to rebuild the appropriate modules
of the stack's logging. This will make it so that when the Unit
Tests are ran, that all logging will actually be off. Until now, the Unit
Tests were re-using the binaries from the last debug build without re-
triggering a rebuild to remove the logging symbols from the CA and RI
Layers. We need the Unit Tests to run with debug symbols to fully enable
the Valgrind pluging on the Jenkins CI System.

Change-Id: I117f583a18be0f8dd9c9f2c52b30b61ceb34ed70
Signed-off-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/687
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>