Fixed Android Build error caused by merge issue
Apparently while merging the Sconscript for the libcoap directory,
I managed to grab some lines from somewhere else(I have no idea
where from!).  This was a mistake, so I'm reverting this file
back to have the identical functionality as the CA branch,
since this doesn't exist on the pre-merge master branch.

Change-Id: I8fcfe692e4e4ab35666ed4a5f02041faffa440c0
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/708
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Rahul Rahul <rahul.rahul@intel.com>