Replace glib threadpool usage with a 'dumb' thread implementation.
We are attempting to remove glib as a necessity for the build, so this
patch replaces the glib threadpool implementation with a 'dumb' dispatch
thread model.

It renames u_threadpool with ca_threadpool to be more internally consistent
with naming conventions.

Finally, it removes all glib build dependencies for non-Tizen build scripts.

Change-Id: Icc64b758ae6c00ff426adda1c74d8f86e56a8fc5
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/747
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-by: Sashi Penta <sashi.kumar.penta@intel.com>